Added exception handling
