--- ButtonTypes.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class ButtonTypes(Enum):
-    SIGN = 1
-    NUMBER = 2
