--- Calculate_Utils.py
@@ -1,7 +1,0 @@
-class Calculate_Utils():
-
-    def sum(x, y):
-        return x + y
-
-    def minus(x, y):
-        return x - y
